requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
diff-match-patch==20241021
Django==4.2.16
django-cors-headers==4.7.0
django-import-export==4.3.14
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
idna==3.10
openpyxl==3.1.5
pillow==12.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
reportlab==4.4.7
requests==2.32.3
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
urllib3==2.4.0
